here's the problem
you flag nodes as dirty for 2 reasons: a child moved, or a
child hash was computed.  Sometimes you get the same node
flagged twice because of these two reasons, so it'll try to
recompute twice.  That's bad.  The obvious solution is a...
map.  Blarg.  There's gotta be a cleaner way to do it.