topDownTransform() is wrong, that's the problem
operating on 7 adds, 5 deletes, the previous bottom up transform gives a result of
6, 5
at the bottom.  Top down does:

topDown on [{10 8} {6 4}]
td output [{10 8} {6 4}]
raw: 10 -> 8	 leaf: 4 -> 0	5 -> 1
6 -> 4	 leaf: 6 -> 4
floor: [{4 0} {5 1} {6 4}]

Which would result in 4, 5 instead of 6, 5.
The problem is that topDown does nothing to the input here, when it should be changing
{6 4} into {6 0}.  4 is under 10 so the 4 should change to a 0 but doesn't.

Fix that, and maybe everything will work!  Or at least a bunch more.