update to add stop / resume
uses a new struct, blockInfo which it saves to disk after every block has
been written.

Some progress is lost in that the readRawBlocksFromFile call has to be
repeated for all the blocks in the file after the one last written.  So on
average, like half a blk___.dat file per stop / resume which is OK.

It seems to work suspending and resuming.  So that's cool.
Still has a bug somewhere where it hits a dead end (reorg block) and doesn't finish.