nope that doesn't work
yeah... that hashable node idea doesn't look like it's going to work
at all.  There can be a move to and from a location in a single row,
and then you have non-deterministic results.

Guess we have to do all swaps on a row, then all hashes.  Which
is the way it worked before.  Blah.