delete a bunch of stuff, and narrow down a problem
the problem seems to be when 2 siblings are deleted, it crashes
trying to hash the parent.  The nodes should be populated if they're
being deleted, so that's not right.  Also we don't even need to hash it
since it's getting deleted anyway, AND nothing's even changed.  So 3 things
are going wrong here.