ok you know what I'm giving up on this pointer based swapNodes
it's way more complicated, swapNodes is already like 80 lines and the
current bug will just make it a bunch more; nodes have to be created in
this case: swap 9, 11 with 9 nil, which wipes out 10's children (11s
nieces) which are still needed.  Can write code for this case, but then
we're up to 100+ lines for something that was 5 lines.
Whatever doing it way simplifies doesn't seem worth it.