remaining issue is with hashing to outside of the forest
generally, you don't need to hash into a location that won't exist, right?
but sometimes that location moves in to somewhere that does exist.
It triggers the "can't hash 0000" panic when it's not moving though.
Really, efficiency wise it'd be nicer to not do useless hashing, so have a
way to detect when it's outside of the forest and not moving in to the forest.
Maybe swaps are sorted in such a way that this is not too expensive to check.