cmd: refactors all files inside cmd/
Everything is refactored.

While reviewing code to do the TODOs mentioned
in #79, I noticed some things to clean up.

This refactor is needed for the goals mentioned
in #79 as it separates

1. Bridge node and CSN (Compact State Node)
2. Chain resume functionality from both bridgenode and csn
(placed in chain.go and chainio.go in both bridgenode/ and csn/

Changed things

1. All functionality for store and restore to disk is now placed in
chain.go and chainio.go in both bridgenode/ and csn/

2. Various bugs involving offsetfile.go was fixed. To add new blocks
from blk*.dat files, all the user has to do is delete the offsetdata
directory and re-run genproofs to rebuild the offset index. Both
genproofs and ibdsim will resume where they left off and sync the
new blocks.

3. main.go file in directory cmd/ is the main file now.

4. Separated all ttl related things to its own directory.

5. Removed all names related or implying a simulator.

6. Various comment cleanups