fix several fun bitcoin specific off-by-1 errors
in the blk.dat and rev.dat files, the block at offset 0 is
block 1.  Fixed that, added a toString method for leaves in
the process.
Still not working due to utxos that get created & destroyed
in the same block;  I guess bridgenode genAddDel() needs to deal
with that, not the accumulator.