updateDirt maintains sortedness
Before this commit, updateDirt would return an unsorted slice,
which would then be passed back into updateDirt. Then,
dedupeSwapDirt would be called with its invariant violated.
So instead of the half-hearted checking of the last element,
we simply do a binary search (cheap) and insert the element
where it should be. No special treating of the last element.