I give up & can't figure it out, will try 1 flat file writer
Having 2 goroutines writing to the same file seems not to work.
I'm trying to seek to somewhere in the middle of the file, and
overwrite 4 bytes.  But that seems to increase the size of the file,
either because the file offset is shared between threads and is writing
in the wrong place, or writes occur out of sync or... something.

Anyway maybe having a single writing worker will be easier as really
we're not getting any concurrency out of this anyway.