the printfs fix the bug...
there are deserialization errors on leafDatas, where it seems
to read the whole leafData in the wrong place, getting outpoint
where the hash has tons of 0s and the outpoint index is way too high,
then it errors on a huge pkScript length.

The errors are different each time, but it dies somewhere in the 400s
(block height) of testnet3.  So it can deal with a few hundred
deserializations before it chokes.

But when the deserializer has printfs that slow it down, it keeps
going fine until the server is done with things to give it.

so... why would timing matter?  I thought TCP was dealing with everything
but I guess it's some buffer empties out when you try to read and some
deserialization part says "OK I'm done" and returns, then the next
deserialize function picks up too early.  Something unexpected anyway.