isolated a problem with op_return
the first op_return output is in block 94766 where it fails.

In the proof.dat file, we get

04ebb527: 0000 0028 0000 0000 0000 0001 0000 0028
The TTLs should be 40, 1, 40, but we're getting 40, 0, 1, 40,
which is probably a space for the op_return which is messing things up.

There's also a 0-block duration spend right next to the op_return output,
but the dedpuing for those seems to work as those are common in many other blocks