there's still a rare case where genproofs skips a block
The proofoffset file had this
00107ab0: 0000 0000 085f 54eb 0000 0000 085f 54eb
where it wrote the same offset twice.  Not sure if it
actually skipped writing the block to proof.dat as well.

Might have something to do with stop / resume; it's not the
same each time and most runs of genproofs do not have this
problem.  I don't think there's any concurrency in
flatFileWorker which is the only other thing I can think of
that would result in this kind of non-deterministic problem