accumulator: Fix batch proof verification (#218)
Before this commit the verification is broken because cached nodes in
the pollard are considered to be the output of two proof hashes.
This is wrong and leads to invalid proofs passing the verification.

The fix is to check whether the cached hash matches the output of the
proof hash. If it matches continue, if it does not match the proof is
invalid and verification should fail.