building ingestAndCheck but need to change grabPos
this will eliminate populate, verifyBatchProof, and ProofPositions,
which is getting rid of a lot of complex code.  So that's good.

It does need a grabPos variant that returns a slice of the whole branch.
Then we can dedupe the nodes in those slices when inserting them into a 2d slice.

modify grabPos, new function, or inline it all into ingestAndCheck..?