accumulator: Change getRootsReverse to getRootsForwards and add a freelist for positionlist (#250)
* accumulator: Change getRootsReverse to getRootsForwards

getRootsReverse forces unnecessary allocations as the roots are
appended to a new list. This helps reduce the amount of memory
allocations, resulting in less time taken up by the gc.

Also, grabbing roots forwards is a lot cleaner than the reverse method.

* accumulator: Use sync.Pool based freelist for positionlist

Positionlist is needed by multiple functions and is allocated and
de-allocated numerous times during the initial block download. Having
a freelist for the positionlist helps reduce gc pressure.