wire: Modify and import the wire package from btcd
The wire package from btcd is imported, removing all unncessary wire
messages for utreexo. The varint encode/decode is removed from the wire
package and is moved to the util package.

This is the first of the commits that aims to internalize necessary utcd
code to the utreexo repo.

bridgenode/accumulator: Implement varint serialization

VarInt serialization that was taken out of wire/ and into util/ is now
used to serialize integers. This saves on bandwidth and saves space on
disk.

csn/: Refactored various parts of the code to fit with the new structure

Some of the pre-existing code from the old wire/ package, that's needed was
moved to the csn/ package. Other refactors are code changes that were
needed to be compatible with the new structure of the utreexo repo.