still errors; after restart txidOffsetFile isn't big enough
TTLLookupWorker tries to read a bit beyond where txidOffsetFile ends.
Either something isn't getting written at the end, or the restoration
process starts too far forward.

Not sure though because stopping the BlockAndRevReader is stopping the
data at the source, and all the data should flow in the same sequence
whether it's interrupted or not.

Maybe the TTL files aren't being flushed properly or something like that...