accumulator/batchproof: Remove target hashes from BatchProof
Before, the target hashes also used to be included in the BatchProof.
However, these are just the hash commitments of elements to prove and
the user has all the data needed to calculate the hashes.

Removing them achieves two things: smaller proofs, simpler verification.

Since the hashes no longer is being stored, the proofs get smaller by
however many deletions that are to happen. The verification logic is
simpler as it no longer has to have extra logic to separate out the
target hashes from the actual proofs.