accumulator/pollard: Fix pollard memory leak
Previously, polNodes that were to be newly allocated would be done so
through a slice. However, since Go gc doesn't garbage collect parts of
slices, most of the polNodes that were created were never garbage
collected. This resulted in the polNodes allocated taking up over 20GBs
on testnet3. This commit reverts the polNode allocation from done all
together as a slice to being done one at a time and not as a slice.