maybe works, but not sure
The problem was I was adding the outpoint index twice: once at the
end of the binSearch function, and then again right after that
returned.  So that's fixed.

But, still two problems:
The bytes read by the binSearch (which is the whole point) are always
0000.  So that can't be right.  And, maybe related, the TTLs are
overwriting themselves; the same location is written more than once.