go.mod, go.sum: Switch to the btcd++ branch in utcd
utcd has many optimizations that btcd doesn't have. Switching to utcd
speeds things up drastically for the binaries in cmd/.

We won't be pointing to the branch utreexo in utcd as breaking changes
to the accumulator package will result in utcd failing to build. This
makes changes annoying to deal with.