csn/ibd: Don't set UTXOs to be remembered
UTXOs are currently marked with ttl value of 0. Since 0 < 1000 is true,
the pollard was remembering all the UTXOs ever, leading to a ton of
memory usage. With this change, CSNs see a lot less memory usage.

This PR still sees the remembered leaves growing over time
so there might still be a problem yet to be fixed.