accumulator/batchproof: Error instead of bool for verifyBatchProof
Now the verifyBatchProof function returns an error instead of a bool to
denote that a proof has failed to validate. If the proof validates, a
nil is returned. If it doesn't, an error with a helpful error message
will be returned.

There may be various reasons for verifyBatchProof to fail but the
message that the caller got was just a bool. Returning an error when the
proof fails allows more information to be tranferred to the caller.