Skiplists are a mess
I see what the problem is.  But not sure how to fix it, so will just write it up for now.

TTL allocation happens for all non-skipped outputs in a block.
The sorted mini-txids also have their "index in block" reference
relative to the skipped locations.  So if there are 2 txs in a block,
and tx 0 (coinbase) has an op_return as output 1 (which all blocks do
by default), then tx 0 will have idxinblock 0, and tx 1 will have
idxinblock 1, not 2.  (because tx 0 only has 1 output that counts)

The problem is if you had, for example, a tx with 3 outputs.  Some
real coins as output 0, op_return as output 1, and real coins as
output 2.  Say that tx's starting position was 10.
Output 0 is position 10 within the block.
Output 1 doesn't exist.
Output 2 is position 11, not 12, within the block.

So in binSearch() when it just adds +mi.idx at the end, that doesn't
work.  We need to also remove the skipped outputs within that tx.
But we don't even have that info anymore.  Seems like we'd have to
keep a skiplist for every previous block somewhere.

It's a mess and I'm sick of flatterttl so I think it's best to just
allocate 4 useless bytes for all the op_returns and other skipped
outputs.  It increases the TTL data by a significant amount, but
TTL data isn't huge.  With better skiplist management that extra
data can be removed, but that can be a TODO for now.