accumulator/pollard: Bug fixes for pollard add.
The siblings of the added leaves will now also be set to be remembered.
This solves a bug where the sibling wasn't present when the proof ingest
doesn't happen before every single modify. This is now tested with the
test function TestPollardNoSiblingFound.

When a leaf hash to be remembered isn't found, the program paniced on a
nil pollard node. There is now a check to make sure the node is not nil
before trying to do operations on it.