more things seem to work
all the deletions that I've tested seem to work the right way (need to test more though)
Deleting a root leaves an empty hash.  In add, if that's detected it needs to not hash and instead promote.
Once that all works, then to work on undo, which seems like it'll be not too tricky